+++ packages/devtools_app/test/shared/diagnostics/inspector_service_test.dart
@@ -127,8 +127,8 @@ void main() {
               await inspectorServiceLocal.isWidgetCreationTracked(),
               isTrue,
             );
-            final rootLibrary =
-                await serviceConnection.mainIsolateRootLibraryUriAsString();
+            final rootLibrary = await serviceConnection.serviceManager
+                .mainIsolateRootLibraryUriAsString();
             await inspectorServiceLocal.addPubRootDirectories([rootLibrary!]);
             final rootDirectories =
                 await inspectorServiceLocal.getPubRootDirectories() ??