+++ packages/devtools_app_shared/lib/src/ui/common.dart
@@ -389,9 +389,7 @@ class DevToolsButton extends StatelessWidget {
               : IconButton(
                   onPressed: onPressed,
                   iconSize: defaultIconSize,
-                  icon: Icon(
-                    icon,
-                  ),
+                  icon: Icon(icon),
                 ),
         ),
       );