+++ packages/devtools_app_shared/CHANGELOG.md
@@ -1,5 +1,10 @@
-## 0.2.0-dev.1
+## 0.2.0
 * Add `navigateToCode` utility method for jumping to code in IDEs.
+* Add `FlutterEvent` and `DeveloperServiceEvent` constants.
+* Add `connectedAppPackageRoot`, `rootPackageDirectoryForMainIsolate`, and
+`mainIsolateRootLibraryUriAsString` methods to the `ServiceManager` class.
+* Bump minimum Dart SDK version to Dart stable `3.4.3` and minimum Flutter SDK
+version to Flutter stable `3.22.2`.
 
 ## 0.2.0-dev.0
 * Add `tooltipWaitExtraLong` to `utils.dart`.